+++ x-pack/plugins/synthetics/public/legacy_uptime/components/overview/integration_deprecation/index.tsx
