+++ x-pack/plugins/synthetics/public/legacy_uptime/pages/overview.tsx
@@ -12,7 +12,7 @@ import styled from 'styled-components';
 import { useTrackPageview } from '@kbn/observability-plugin/public';
 import { useBreadcrumbs } from '../hooks/use_breadcrumbs';
 import { MonitorList } from '../components/overview/monitor_list/monitor_list_container';
-import { IntegrationDeprecation } from '../components/overview/zip_url_deprecation';
+import { IntegrationDeprecation } from '../components/overview/integration_deprecation';
 import { StatusPanel } from '../components/overview/status_panel';
 import { QueryBar } from '../components/overview/query_bar/query_bar';
 import { MONITORING_OVERVIEW_LABEL } from '../routes';
